[tests] approval testing Security - Replaced dangerous Path.Combine with Path.Join

+++ Mapp.BusinessLogic.Transactions/Transaction.cs
@@ -20,6 +20,6 @@ namespace Mapp.BusinessLogic.Transactions
 
         public string OrderId { get; set; }
 
-        public double TransactionValue { get; set; }
+        public decimal TransactionValue { get; set; }
     }
 }
